requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
av==16.0.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
ctranslate2==4.6.2
distro==1.9.0
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.2
flatbuffers==25.12.19
frida==17.5.2
frida-tools==14.5.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
imageio-ffmpeg==0.6.0
jiter==0.12.0
mpmath==1.3.0
numpy==2.4.0
onnxruntime==1.23.2
openai==2.14.0
packaging==25.0
prompt_toolkit==3.0.52
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==13.1
